+++ cosmictiger/memory.hpp
@@ -65,7 +65,7 @@ void cuda_unified_show_outstanding();
 #define FREE(ptr) free(ptr)
 
 template<class T>
-CUDA_EXPORT inline void cuda_malloc(T **ptr, int nele, const char *file, int line) {
+CUDA_EXPORT inline void cuda_malloc(T **ptr, size_t nele, const char *file, int line) {
 #ifdef __CUDA_ARCH__
 	*ptr = (T*) malloc(nele * sizeof(T));
 	MEM_CHECK_POINTER(*ptr, file, line);
