+++ src/math.cu
@@ -88,7 +88,7 @@ void generate_random_vectors(fixed32* x, fixed32* y, fixed32* z, size_t N, int s
 	__syncthreads();
 }
 
-void generate_random_normals(cmplx* rands, int N, int seed) {
+void generate_random_normals(cmplx* rands, size_t N, int seed) {
 	generate_random_normals_kernel<<<32,32>>>(rands, N, seed);
 	CUDA_CHECK(cudaDeviceSynchronize());
 }
